[Ref] examples: Silence warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@8688 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ examples/libopenmpt_example_c_probe.c
@@ -171,7 +171,7 @@ static blob_t * load_file( const char * filename ) {
 	if ( tell_result < 0 ) {
 		goto fail;
 	}
-	if ( (unsigned long)tell_result > SIZE_MAX ) {
+	if ( (unsigned long long)tell_result > SIZE_MAX ) {
 		goto fail;
 	}
 	blob->size = (size_t)tell_result;
